feat: add empty folder warning/variable and verbose option
This commit adds the EMPTY_FOLDER environment variable which can be used
to check if the folder that was specified was empty. It now also throws
a warning in the github action console when no files are found. Lastly,
the `verbose` input argument was added which allows people to print the
black output to the github action console. It also adds a test to see if
the empty folder check works.