fix: fix 'bad array subscript' error (#62)
* fix: fix 'bad array subscript' error

This commit fixes the 'bad array subscript' error that is thrown when no
python files are found (see https://github.com/rickstaa/panda-gazebo/actions/runs/5269309929/jobs/9527260724).
It also simplifies the codebase.

* feat: add empty folder warning/variable and verbose option

This commit adds the EMPTY_FOLDER environment variable which can be used
to check if the folder that was specified was empty. It now also throws
a warning in the github action console when no files are found. Lastly,
the `verbose` input argument was added which allows people to print the
black output to the github action console. It also adds a test to see if
the empty folder check works.

* fix: fix verbose input bug

* fix: upload empty folder

* test: fix verbose test

* feat: remove EMPTY_FOLDER env variable

* fix: fix warning color

* fix: uncomment code

* fix: remove test code